asdf

+++ Emvco/Emv/Common/Play.Emv.Interchange/Exceptions/Check.cs
@@ -20,7 +20,7 @@ internal class Check
         /// <param name="length"></param>
         /// <param name="dataFieldId"></param>
         /// <exception cref="InterchangeDataFieldOutOfRangeException"></exception>
-        public static void ForExactLength<T>(T[] value, int length, DataFieldId dataFieldId) where T : struct
+        public static void ForExactLength<T>(T[] value, nint length, DataFieldId dataFieldId) where T : struct
         {
             if (value.Length != length)
             {
@@ -34,7 +34,7 @@ internal class Check
         /// <param name="length"></param>
         /// <param name="dataFieldId"></param>
         /// <exception cref="InterchangeDataFieldOutOfRangeException"></exception>
-        public static void ForExactLength<T>(ReadOnlySpan<T> value, int length, DataFieldId dataFieldId) where T : struct
+        public static void ForExactLength<T>(ReadOnlySpan<T> value, nint length, DataFieldId dataFieldId) where T : struct
         {
             if (value.Length != length)
             {
